reset nav plans
